delete key handler on LineItem doesn't remove non-empty parents anymore

+++ systori/dart/web/editor.dart
@@ -968,9 +968,9 @@ class LineItemElement extends EditableElement {
 
     void handle_delete_key(event) {
         var saved_parent = this.parent;
-
         super.handle_delete_key(event);
-        if (saved_parent.can_delete()) {
+        
+        if (event.shiftKey && saved_parent.can_delete() && saved_parent.classes.contains('empty')) {
             saved_parent.delete();
             saved_parent.remove();
         }
